failing test case for different locals w/ same template proc
The problem here is that the cached template proc closes over the
locals variable when the proc is created and not when the proc is
instance eval'd. Caching the proc on the entire locals Hash (instead
of just the keys) fixes this but holds references to all locals
values and has really poor cache hit characteristics.