revise and define more strictly the default_encoding option
The default_encoding option takes effect only when tilt reads
template data from the filesystem. Templates provided via custom
reader block are assumed to be tagged with a best guess encoding
already.

It's also worth noting that, unlike File.read, the default file
reader does not perform Encoding.default_internal transcoding. The
string is marked with the default_encoding or the system encoding
(Encoding.default_external) but no transcoding is performed. This is
because magic comments or template specific encoding settings are
not yet available.