Preload files instead of autoload if the engine class has been defined.
A final argument is added to register_lazy, an optional options hash.
There is one option, :preload_if.  Its value is a symbol or an Array of
symbols.  When new method Tile.preload is called, if any of the symbols
(as strings) is a defined constant then the associated file (from
register_lazy) is required immediately instead of waiting to be autoloaded.

This prevents the "non thread-safe" autoload warning without requiring
tilt/haml, etc., to be explicitly required.  All that's needed is for the
engines to be required before Tilt.preload is called:

  require "haml"
  require "redcarpet"
  require "tilt"
  ...
  Tilt.preload

Note that in general tilt can't do the preload itself because haml (at least)
does its own "require 'tilt'".

Calling Tilt.preload is optional.