Make #render handle nil scope arg consistently
Docs (e.g. `docs/TEMPLATES.md`) appear to show that `nil` can be
passed as scope argument to `#render` and that Tilt will do the
"right thing". Some real-world usages also have this expectation.

    Tilt::HamlTemplate.new('example.haml')
    t.render(nil, local_var: "testing")

In the case of `Tilt::HamlTemplate` the above actually results in the
`NilClass` singleton being used as scope, and in turn ivars being set
on that same object. In versions of Ruby prior to 2.2, this was not a
visible issue. As of Ruby 2.2, where `NilClass` is a frozen object,
the above will fail.

Fixes #258

Signed-off-by: Alex Coles <alex@alexbcoles.com>