Add Tilt::Template#freeze_string_literals?
If this returns true, frozen-string-literal: true is set when
compiling the template method.  It is false by default, and
will need to be manually set to true in the template classes
that could support it.

Support :freeze option for ERB, Erubis, and Erubi template to
use frozen string literals.

Avoid printing deprecation warnings when running Erubis tests.
Similar change already made when running the ERB tests.