Handle rendering BasicObject instances
There are a couple of issues here.  One is that BasicObjects do
not respond to is_a?.  Another is that they don't respond to
class.

To handle the class issue, use a case statement to separate
BasicObject from Object, and bind a Kernel method to the BasicObject
instance to be able to get the class.  Note that this does not work
on Ruby 1.9, and on Ruby 1.9, there is not a way to correctly get
the scope class.  Use Object in that case so things don't completely
break (though constant lookup will not use use the scope class).

While we don't need to handle is_a? after that change, still switch
from scope.is_a?(Module) to Module === scope, as that handles cases
where is_a? is redefined.