Remove use of Bundler.setup
If use of bundler is desired, bundle exec should be used.  This
is already done in CI.

With the current setup, I cannot even run rake -T to see available
tasks without installing gems.  When running tests for one engine,
I should not need to install tests for other engines I'm not using.

All template engine tests already have checks for whether the
engine is installed, and if the engine is not available, it prints
a useful message and skips the tests for that engine.