Fix incorrect test
The "encoding with magic comment" test was supposed to check
that Tilt will respect the magic comment of the template source.
However, Tilt doesn't do that. Tilt will extract the magic comment
of generated ruby code for precompiled templates.  However, the
template source code must be valid in the default external encoding
if you do not provide a :default_encoding option.

The test worked correctly when the default external encoding was
compatible with the magic comment, but it didn't even
check that the encoding of the resulting render matched.  Modify
the test to check that it will use the source encoding if given.

Also modify the following test to ensure it actually tests that
render result has the correct encoding.