Remove more Ruby < 2 compatibility code
Split TiltTemplateTest into two classes. There is setup/teardown
blocks in the middle of the file that you might expect are only
for the following tests, but they actually apply to all tests
in the file, including the tests above the setup/teardown blocks
that aren't using what the setup/teardown blocks provide.

This is best viewed ignoring whitespace differences.