Fix Gemfile to get more template engines tested in CI
The erubi template was not tested in CI as it wasn't in the Gemfile.
Add erubi to Gemfile in the primary group, since it has priority for
.erb files, and move erubis to the secondary group.

The less gem was installed in the primary group, but the tests
wouldn't run as therubyracer gem wasn't installed.  Move less to
the secondary group, and install therubyracer.

Remove the restriction on the pdf-reader gem, since tests pass with
the current version of pdf-reader. That should hopefully allow the
prawn tests to run in CI (they previously didn't).

After this, the only tests not run in CI are BlueCloth and Sigil.
We may want to deprecate those in 2.1 and remove in 2.2.