Deprecate BlueCloth, Less, and Sigil support
None of these are tested in CI.

BlueCloth conflicts with RDiscount.  It could theoretically be
tested in CI if RDiscount was in primary templates and BlueCloth
in secondary templates, but considering BlueCloth hasn't had a
release in over 10 years, and Tilt supports many other markdown
formats, there isn't a good reason to keep supporting it.

Less on CRuby depends on therubyracer, which no longer works on
modern Ruby due to changes in Psych.  I tried enabling it in CI,
and it won't install (https://github.com/jeremyevans/tilt/runs/7812514562?check_suite_focus=true#step:4:155).
Less hasn't had a release in over 8 years, and therubyracer hasn't
had a release in over 5 years.

Sigil doesn't have a gem, and isn't even packaged for my operating
system (OpenBSD).  It's possible we could test it in CI if it is
packaged for whatever Linux distro CI uses, but I doubt the usage
of tilt/sigil is high enough to justify that.

If any tilt user really needs these template engines, they can just
copy the related template file and require it manually.